--- screens/orders/OrdersScreen.js
@@ -1,20 +1,0 @@
-import React from 'react';
-import { StyleSheet, View, Text } from 'react-native';
-import GlobalStyles from '../../config/styles';
-
-const OrdersScreen = () => {
-    return (
-        <View style={style.container}>
-            <Text>OrdersScreen</Text>
-        </View>
-    )
-}
-
-const style = StyleSheet.create({
-    container: {
-        backgroundColor: GlobalStyles.color.primary,
-        height: "100%"
-    }
-});
-
-export default OrdersScreen;
